Added PeriodData class which will server as our main Datastore for measurements. Edited Calculations class with a slight error in windChill Edited Measurement class to return sunrise and sunset in doubles. Edited Period class with new comment, changing Johan's one. Added init function to Period class for automization. Added setData function which does most calculations and sets them in the PeriodData datastorage (example of how to call it in main) Note: Apart from Mist and RisingTemperature, the current individual assignments DO NOT WORK due to incompatable input types.

+++ src/Calculations.java
@@ -102,7 +102,7 @@ public class Calculations {
         return result;
     }
 
-    public static double windChill(short outsideTemp, short windSpeed) {
+    public static double windChill(double outsideTemp, double windSpeed) {
         double windchill = 13.12 + (0.6215 * windSpeed) - (11.37 * Math.pow(outsideTemp, 0.16)) + ((0.3965 * windSpeed) * Math.pow(outsideTemp, 0.16));
         return windchill;
     }
